+++ android/pad/app/src/main/java/eu/id3/face/samples/pad/CameraFragment.kt
@@ -5,7 +5,6 @@ import android.content.Context
 import android.content.pm.PackageManager
 import android.content.res.Configuration
 import android.graphics.*
-import android.graphics.ImageFormat
 import android.hardware.camera2.*
 import android.hardware.camera2.params.OutputConfiguration
 import android.hardware.camera2.params.SessionConfiguration
@@ -19,7 +18,6 @@ import android.util.Size
 import android.view.*
 import androidx.core.app.ActivityCompat
 import androidx.fragment.app.Fragment
-import eu.id3.face.*
 import java.util.*
 import kotlin.system.exitProcess
 
@@ -61,24 +59,24 @@ class CameraFragment : Fragment() {
         }
     private var cameraId: String = ""
     var cameraDevice: CameraDevice? = null
-    private val cameraDeviceStateCallback: CameraDevice.StateCallback = object : CameraDevice.StateCallback() {
-        override fun onOpened(camera: CameraDevice) {
-            //This is called when the camera is open
-            Log.e(LOG_TAG, "onOpened")
-            cameraDevice = camera
-            createCameraPreview()
-        }
+    private val cameraDeviceStateCallback: CameraDevice.StateCallback =
+        object : CameraDevice.StateCallback() {
+            override fun onOpened(camera: CameraDevice) {
+                //This is called when the camera is open
+                cameraDevice = camera
+                createCameraPreview()
+            }
 
-        override fun onDisconnected(camera: CameraDevice) {
-            cameraDevice?.close()
-            cameraDevice = null
-        }
+            override fun onDisconnected(camera: CameraDevice) {
+                cameraDevice?.close()
+                cameraDevice = null
+            }
 
-        override fun onError(camera: CameraDevice, error: Int) {
-            cameraDevice?.close()
-            cameraDevice = null
+            override fun onError(camera: CameraDevice, error: Int) {
+                cameraDevice?.close()
+                cameraDevice = null
+            }
         }
-    }
     private var cameraCaptureSessions: CameraCaptureSession? = null
     private var captureRequestBuilder: CaptureRequest.Builder? = null
     private var imageDimension: Size? = null
@@ -89,6 +87,8 @@ class CameraFragment : Fragment() {
     private var displayOrientation = 0
     private var isCapturing = false
     private var needsToProcess = false
+    private var needsToEnrollTemplate = false
+    private var needsToVerifyTemplate = false
     private var faceProcessor: FaceProcessor? = null
     private var faceProcessorListener: FaceProcessorListener? = null
 
@@ -125,6 +125,11 @@ class CameraFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
         cameraPreview = view.findViewById(R.id.cameraPreview)
         boundsView = view.findViewById(R.id.boundsView)
+        val boundsPaint = Paint(Paint.ANTI_ALIAS_FLAG)
+        boundsPaint.color = Color.GREEN
+        boundsPaint.strokeWidth = 3.toFloat()
+        boundsPaint.style = Paint.Style.STROKE
+        boundsView.setPaint(boundsPaint)
         boundsView.setWillNotDraw(false)
         boundsView.setLayerType(View.LAYER_TYPE_SOFTWARE, null)
     }
@@ -141,6 +146,22 @@ class CameraFragment : Fragment() {
      */
     fun setProcessor(processor: FaceProcessor) {
         faceProcessor = processor
+    }
+
+    /**
+     * Requires the camera background thread to enroll a face template.
+     * It triggers the enrolment operation later in the 'onImageAvailableListener'.
+     */
+    fun requestTemplateEnrolment() {
+        needsToEnrollTemplate = true
+    }
+
+    /**
+     * Requires the camera background thread to verify a face template.
+     * It triggers the verification operation later in the 'onImageAvailableListener'.
+     */
+    fun requestTemplateVerification() {
+        needsToVerifyTemplate = true
     }
 
     /**
@@ -196,7 +217,6 @@ class CameraFragment : Fragment() {
      */
     override fun onResume() {
         super.onResume()
-        Log.v(LOG_TAG, "onResume")
         startBackgroundThread()
         if (cameraPreview.isAvailable) {
             openCamera(cameraPreview.width, cameraPreview.height)
@@ -209,7 +229,6 @@ class CameraFragment : Fragment() {
      * Pauses the camera preview.
      */
     override fun onPause() {
-        Log.v(LOG_TAG, "onPause")
         closeCamera()
         stopBackgroundThread()
         super.onPause()
@@ -280,7 +299,7 @@ class CameraFragment : Fragment() {
             for (camId in manager.cameraIdList) {
                 val characteristics = manager.getCameraCharacteristics(camId)
                 val cOrientation = characteristics.get(CameraCharacteristics.LENS_FACING)
-                if (cOrientation == CameraCharacteristics.LENS_FACING_FRONT) {
+                if (cOrientation == Parameters.cameraType) {
                     cameraId = camId
                     break
                 }
@@ -376,15 +395,11 @@ class CameraFragment : Fragment() {
             if (ActivityCompat.checkSelfPermission(
                     this.requireContext(),
                     Manifest.permission.CAMERA
-                ) != PackageManager.PERMISSION_GRANTED &&
-                ActivityCompat.checkSelfPermission(
-                    this.requireContext(),
-                    Manifest.permission.WRITE_EXTERNAL_STORAGE
                 ) != PackageManager.PERMISSION_GRANTED
             ) {
                 ActivityCompat.requestPermissions(
                     this.requireActivity(),
-                    arrayOf(Manifest.permission.CAMERA, Manifest.permission.WRITE_EXTERNAL_STORAGE),
+                    arrayOf(Manifest.permission.CAMERA),
                     REQUEST_CAMERA_PERMISSION
                 )
                 return
@@ -402,8 +417,10 @@ class CameraFragment : Fragment() {
      * Closes the camera.
      */
     private fun closeCamera() {
-        cameraDevice!!.close()
-        imageReader!!.close()
+        if (cameraDevice != null)
+            cameraDevice!!.close()
+        if (imageReader != null)
+            imageReader!!.close()
     }
 
     /**
@@ -449,11 +466,16 @@ class CameraFragment : Fragment() {
              * Resize image if larger or higher than 'maxProcessingImageSize'.
              * This operation allows to speed up the face detection process.
              */
-            processingImage.downscale(Parameters.maxProcessingImageSize)
+            val downscaledImage = processingImage.clone()
+            val downscaleRatio = downscaledImage.downscale(Parameters.maxProcessingImageSize)
 
             /** Track faces. */
-            val detectedFace = faceProcessor?.trackLargestFace(processingImage)
+            val detectedFace = faceProcessor?.detectLargestFace(downscaledImage)
             if (detectedFace != null) {
+                /*
+                 * For PAD operations it is better to work on the original image so we rescale the detectedFace.
+                 */
+                detectedFace.rescale(1 / downscaleRatio)
                 val bounds = detectedFace.bounds
 
                 /**
@@ -470,7 +492,8 @@ class CameraFragment : Fragment() {
 
                 /** Process frame if requested by UI. */
                 if (needsToProcess) {
-                    val analyzeLargestFaceResult = faceProcessor!!.analyzeLargestFace(processingImage, detectedFace)
+                    val analyzeLargestFaceResult =
+                        faceProcessor!!.analyzeLargestFace(processingImage, detectedFace)
                     faceProcessorListener!!.onLargestFaceProcessed(analyzeLargestFaceResult)
                     needsToProcess = false
                 }