+++ android/recognition-java/app/src/main/java/eu/id3/face/samples/recognitionjava/FaceCapturePreviewView.java
@@ -1,0 +1,49 @@
+package eu.id3.face.samples.recognitionjava;
+
+import android.content.Context;
+import android.util.AttributeSet;
+import android.view.TextureView;
+
+import androidx.annotation.NonNull;
+
+public class FaceCapturePreviewView extends TextureView {
+    private int ratioWidth = 0;
+    private int ratioHeight = 0;
+
+    public FaceCapturePreviewView(@NonNull Context context) {
+        super(context);
+    }
+
+    public FaceCapturePreviewView(Context context, AttributeSet attrs) {
+        super(context, attrs);
+    }
+
+    public FaceCapturePreviewView(Context context, AttributeSet attrs, int defStyle) {
+        super(context, attrs, defStyle);
+    }
+
+    protected void setAspectRatio(int width, int height) {
+        if (width < 0 || height < 0) {
+            throw new IllegalArgumentException("Size cannot be negative.");
+        }
+        ratioWidth = width;
+        ratioHeight = height;
+        requestLayout();
+    }
+
+    @Override
+    protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
+        super.onMeasure(widthMeasureSpec, heightMeasureSpec);
+        int width = MeasureSpec.getSize(widthMeasureSpec);
+        int height = MeasureSpec.getSize(heightMeasureSpec);
+        if (0 == ratioWidth || 0 == ratioHeight) {
+            setMeasuredDimension(width, height);
+        } else {
+            if (width < height * ratioWidth / ratioHeight) {
+                setMeasuredDimension(width, width * ratioHeight / ratioWidth);
+            } else {
+                setMeasuredDimension(height * ratioWidth / ratioHeight, height);
+            }
+        }
+    }
+}