+++ android/recognition/app/src/main/java/eu/id3/face/samples/recognition/Parameters.kt
@@ -1,6 +1,7 @@
 package eu.id3.face.samples.recognition
 
-import eu.id3.face.*
+import android.hardware.camera2.CameraCharacteristics
+import eu.id3.face.FaceMatcherThreshold
 
 /**
  * Contains the parameters needed in the app. Values of these parameters can be changed and should not be seen as "the perfect" values.
@@ -12,4 +13,5 @@ internal object Parameters {
     const val encodingQualityThreshold = 40
     val fmrThreshold = FaceMatcherThreshold.FMR10000
     const val maxProcessingImageSize = 512
+    const val cameraType = CameraCharacteristics.LENS_FACING_FRONT
 }