+++ flutter/pad_sample/lib/constants.dart
@@ -1,3 +1,2 @@
 const blurScoreMaxThreshold = 20;
 const colorScoreThreshold = 90;
-const moireScoreThreshold = 10;